Move componenets to mart.utils.config.
