+++ mart/generate_config.py
@@ -6,47 +6,15 @@
 
 from __future__ import annotations
 
-import os
-
 import fire
-from hydra import compose, initialize_config_dir
-from omegaconf import OmegaConf
 
-DEFAULT_VERSION_BASE = "1.2"
-DEFAULT_CONFIG_DIR = "."
-DEFAULT_CONFIG_NAME = "lightning.yaml"
-
-
-def mart_compose(
-    *overrides,
-    version_base: str = DEFAULT_VERSION_BASE,
-    config_dir: str = DEFAULT_CONFIG_DIR,
-    config_name: str = DEFAULT_CONFIG_NAME,
-    export_node: str | None = None,
-):
-    # Add an absolute path {config_dir} to the search path of configs, preceding those in mart.configs.
-    if not os.path.isabs(config_dir):
-        config_dir = os.path.abspath(config_dir)
-
-    # hydra.initialize_config_dir() requires an absolute path,
-    # while hydra.initialize() searches paths relatively to mart.
-    with initialize_config_dir(version_base=version_base, config_dir=config_dir):
-        cfg = compose(config_name=config_name, overrides=overrides)
-
-    # Export a sub-tree.
-    if export_node is not None:
-        for key in export_node.split("."):
-            cfg = cfg[key]
-
-    return cfg
-
-
-def get_yaml_cfg(cfg, resolve: bool = False):
-    # Resolve all interpolation in the sub-tree.
-    if resolve:
-        OmegaConf.resolve(cfg)
-
-    return OmegaConf.to_yaml(cfg)
+from .utils.config import (
+    DEFAULT_CONFIG_DIR,
+    DEFAULT_CONFIG_NAME,
+    DEFAULT_VERSION_BASE,
+    compose,
+    get_yaml_cfg,
+)
 
 
 def main(
@@ -57,7 +25,7 @@ def main(
     export_node: str | None = None,
     resolve: bool = False,
 ):
-    cfg = mart_compose(
+    cfg = compose(
         *overrides,
         version_base=version_base,
         config_dir=config_dir,
